feat: create signUpScreen.js
